pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "youtube-suite"
version = "0.1.0"
description = "Unified YouTube market intelligence + creator studio (subs, shorts)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.104",
  "uvicorn[standard]>=0.24",
  "sqlalchemy>=2.0",
  "psycopg2-binary>=2.9",
  "alembic>=1.13",
  "pydantic>=2.5",
  "pydantic-settings>=2.0",
  "python-dotenv>=1.0",
  "typer>=0.12",
  "google-api-python-client>=2.100",
  "python-dateutil>=2.8",
  "click>=8.0",
  "openai>=1.0",
  "ffmpeg-python>=0.2.0",
  "faster-whisper>=1.0",
  "python-multipart>=0.0.6",
]

[project.optional-dependencies]
shorts = [
  "openai-whisper",
  "whisperx==3.1.1",
  "torch>=2.0",
  "torchaudio>=2.0",
  "pyannote.audio==3.1.1",
  "sentence-transformers==2.2.2",
  "numpy>=1.24",
  "scipy>=1.10",
]
dev = [
  "pytest>=7.4",
  "httpx>=0.25",
]

[project.scripts]
cip = "youtube_suite.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
